requirements.txt
django==4.2
django-rest-framework=3.15.1

# repostat
Jinja2>=2.10.1
pygit2>=1.6.1
pytz>=2018.5
pandas>=1.3.3
tqdm~=4.45.0

